requirements.txt
requests
shodan
pyyaml
